11_14_class/mobile-payment-app/src/mobile_payment_app/services/qr_scanner.py
from flask import jsonify
import qrcode
import cv2

def generate_qr_code(data):
    qr = qrcode.QRCode(version=1, box_size=10, border=5)
    qr.add_data(data)
    qr.make(fit=True)
    img = qr.make_image(fill='black', back_color='white')
    return img

def scan_qr_code(image_path):
    img = cv2.imread(image_path)
    detector = cv2.QRCodeDetector()
    data, bbox, _ = detector(img)
    if bbox is not None:
        return data
    return None